+++ user-api/src/main/java/com/biengual/userapi/learning/domain/UserLearningHistoryCustomRepository.java
@@ -10,6 +10,7 @@ import com.querydsl.jpa.impl.JPAQueryFactory;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 
@@ -23,7 +24,7 @@ public class UserLearningHistoryCustomRepository {
     private final JPAQueryFactory queryFactory;
 
     // 해당 컨텐츠의 유저 학습률을 조회하기 위한 쿼리
-    public Optional<Integer> findLearningRateByUserIdAndContentId(Long userId, Long contentId) {
+    public Optional<BigDecimal> findLearningRateByUserIdAndContentId(Long userId, Long contentId) {
         return Optional.ofNullable(
             queryFactory
             .select(userLearningHistoryEntity.learningRate)
